Remove validations endpoints
Validations will be stored in HDFS and cannot be queried by the data API.
