--- config/validators.config.json.example
@@ -1,8 +1,0 @@
-{
-  "validator-domains" : {
-    "n9LdgEtkmGB9E2h3K4Vp7iGUaKuq23Zr32ehxiU8FWY7xoxbWTSA" : "ripple.com"
-  },
-  "unverified-validators" : [
-    "n9KtJcYZZ9pSJ924u1Pc9F2P1H2oUyR6JYjToMcHrH7Y18GSh6ao"
-  ]
-}
