+++ import/createTables.js
@@ -4,4 +4,7 @@ var rest = new Rest(config.get('hbase-rest'));
 
 rest.initTables('ledgers', function(err, resp) {
   console.log(err, resp);
+  rest.initTables('validators', function(err, resp) {
+    console.log(err, resp);
+  });
 });
