+++ lib/geolocation.js
@@ -2,8 +2,8 @@
 
 const request = require('request-promise');
 const Promise = require('bluebird');
-const hbase = require('../hbase');
-const Logger = require('../logger');
+const hbase = require('./hbase');
+const Logger = require('./logger');
 const log = new Logger({scope: 'geolocation'});
 
 module.exports = config => {
