+++ lib/hbase/hbase-rest.js
@@ -50,10 +50,7 @@ var tables = {
     'payment_channels',
     'account_payment_channels'
   ],
-  validations: [
-    'validations_by_ledger',
-    'validations_by_validator',
-    'validations_by_date',
+  validators: [
     'validator_reports',
     'validators',
     'validator_domain_changes',
@@ -81,7 +78,7 @@ function Client(options) {
     var self = this
 
     Promise.map(tables[group], function(table) {
-      return self.addTable(table, group === 'validations')
+      return self.addTable(table, group === 'validators')
     })
     .nodeify(function(err, resp) {
