+++ lib/hbase/hbase-thrift/topology.js
@@ -473,117 +473,6 @@ function getValidators(options) {
   }
 }
 
-/**
- * getLedgerValidations
- */
-
-function getLedgerValidations(options) {
-  var self = this
-  return new Promise(function(resolve, reject) {
-    self.getScanWithMarker(self, {
-      table: 'validations_by_ledger',
-      startRow: options.ledger_hash,
-      stopRow: options.ledger_hash + '~',
-      marker: options.marker,
-      limit: options.limit,
-      descending: false
-    },
-    function(err, resp) {
-      if (err) {
-        reject(err)
-
-      } else {
-        resp.rows.forEach(function(r) {
-          if (r.count) {
-            r.count = Number(r.count)
-          }
-
-          delete r.rowkey
-        })
-
-        resolve(resp)
-      }
-    })
-  })
-}
-
-
-/**
- * getValidations
- */
-
-function getValidations(options) {
-  var self = this
-  var table
-  var startRow
-  var stopRow
-
-  if (options.pubkey) {
-    table = 'validations_by_validator'
-    startRow = options.pubkey + '|' + options.start.hbaseFormatStartRow()
-    stopRow = options.pubkey + '|' + options.end.hbaseFormatStopRow()
-
-  } else {
-    table = 'validations_by_date'
-    startRow = options.start.hbaseFormatStartRow()
-    stopRow = options.end.hbaseFormatStopRow()
-  }
-
-  return new Promise(function(resolve, reject) {
-    self.getScanWithMarker(self, {
-      table: table,
-      startRow: startRow,
-      stopRow: stopRow,
-      marker: options.marker,
-      limit: options.limit,
-      descending: options.descending,
-      filterString: 'KeyOnlyFilter()'
-    },
-    function(err, resp) {
-      if (err) {
-        reject(err)
-
-      } else {
-        var keys = []
-        resp.rows.forEach(function(r) {
-          var parts = r.rowkey.split('|')
-          if (options.pubkey) {
-            keys.push([
-              parts[2],
-              parts[0],
-              parts[1].substr(0, 14)
-            ].join('|'))
-
-          } else {
-            keys.push([
-              parts[2],
-              parts[1],
-              parts[0].substr(0, 14)
-            ].join('|'))
-          }
-        })
-
-        self.getRows({
-          table: 'validations_by_ledger',
-          rowkeys: keys
-        }, function(err2, rows) {
-          if (err2) {
-            reject(err2)
-
-          } else {
-            rows.forEach(function(r) {
-              r.count = Number(r.count)
-              delete r.rowkey
-            })
-
-            resp.rows = rows
-            resolve(resp)
-          }
-        })
-      }
-    })
-  })
-}
 
 function getNodeState(options) {
   return this.getAllRows({
@@ -598,7 +487,5 @@ module.exports = {
   getTopologyInfo: getTopologyInfo,
   getTopologyLinks: getTopologyLinks,
   getValidatorReports: getValidatorReports,
-  getLedgerValidations: getLedgerValidations,
-  getValidators: getValidators,
-  getValidations: getValidations
+  getValidators: getValidators
 }
