+++ test/mock/validator-reports.json
@@ -108,5 +108,50 @@
     "other_ledgers": "18",
     "total_ledgers": "22476",
     "validation_public_key": "n9Kk6U5nSF8EggfmTpMdna96UuXWAVwSsDSXRkXeZ5vLcAFk77tr"
+  },
+  { "validation_public_key": "n9KcuH7Y4q4SD3KoS5uXLhcDVvexpnYkwciCbcX131ehM5ek2BB6",
+    "date": "2017-10-05T00:00:00Z",
+    "total_ledgers": 6,
+    "main_net_ledgers": 0,
+    "alt_net_ledgers": 0,
+    "other_ledgers": 6,
+    "main_net_agreement": "0.00000",
+    "alt_net_agreement": "0.00000"
+  },
+  { "validation_public_key": "n9Kk6U5nSF8EggfmTpMdna96UuXWAVwSsDSXRkXeZ5vLcAFk77tr",
+    "date": "2017-10-05T00:00:00Z",
+    "total_ledgers": 6,
+    "main_net_ledgers": 0,
+    "alt_net_ledgers": 0,
+    "other_ledgers": 6,
+    "main_net_agreement": "0.00000",
+    "alt_net_agreement": "0.00000"
+  },
+  { "validation_public_key": "n9LXZBs2aBiNsgBkhVJJjDX4xA4DoEBLycF6q8zRhXD1Zu3Kwbe4",
+    "date": "2017-10-05T00:00:00Z",
+    "total_ledgers": 5,
+    "main_net_ledgers": 0,
+    "alt_net_ledgers": 0,
+    "other_ledgers": 5,
+    "main_net_agreement": "0.00000",
+    "alt_net_agreement": "0.00000"
+  },
+  { "validation_public_key": "n9LYyd8eUVd54NQQWPAJRFPM1bghJjaf1rkdji2haF4zVjeAPjT2",
+    "date": "2017-10-05T00:00:00Z",
+    "total_ledgers": 6,
+    "main_net_ledgers": 0,
+    "alt_net_ledgers": 0,
+    "other_ledgers": 6,
+    "main_net_agreement": "0.00000",
+    "alt_net_agreement": "0.00000"
+  },
+  { "validation_public_key": "n9MnXUt5Qcx3BuBYKJfS4fqSohgkT79NGjXnZeD9joKvP3A5RNGm",
+    "date": "2017-10-05T00:00:00Z",
+    "total_ledgers": 5,
+    "main_net_ledgers": 0,
+    "alt_net_ledgers": 0,
+    "other_ledgers": 5,
+    "main_net_agreement": "0.00000",
+    "alt_net_agreement": "0.00000"
   }
 ]
