+++ test/test.healthCheck.js
@@ -162,45 +162,3 @@ describe('health check - Nodes ETL', function() {
     });
   });
 });
-
-describe('health check - Validations ETL', function() {
-  it('should check health', function(done) {
-    request({
-      url: baseURL + '/validations_etl'
-    },
-    function(err, res, body) {
-      assert.ifError(err);
-      assert.strictEqual(res.statusCode, 200);
-      assert.strictEqual(body, '1');
-      done();
-    });
-  });
-
-  it('should check health (verbose)', function(done) {
-    request({
-      url: baseURL + '/validations_etl?verbose=true',
-      json: true
-    },
-    function(err, res, body) {
-      assert.ifError(err);
-      assert.strictEqual(res.statusCode, 200);
-      assert.strictEqual(body.score, 1);
-      assert.strictEqual(body.message, 'last imported data exceeds threshold');
-      done();
-    });
-  });
-
-  it('should use custom threshold', function(done) {
-    request({
-      url: baseURL + '/validations_etl?verbose=true&threshold=Infinity',
-      json: true
-    },
-    function(err, res, body) {
-      assert.ifError(err);
-      assert.strictEqual(res.statusCode, 200);
-      assert.strictEqual(body.score, 0);
-      assert.strictEqual(body.gap_threshold, 'Infinity');
-      done();
-    });
-  });
-});
