--- test/validators.config.json
@@ -1,8 +1,0 @@
-{
-  "validator-domains": {
-    "n9LYyd8eUVd54NQQWPAJRFPM1bghJjaf1rkdji2haF4zVjeAPjT2": "ripple.com"
-  },
-  "unverified-validators": [
-    "n9Kk6U5nSF8EggfmTpMdna96UuXWAVwSsDSXRkXeZ5vLcAFk77tr"
-  ]
-}
